requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.24.0
pillow>=9.0.0